dolt_watchdog: decode processlist.time as i64 (signed INT, not unsigned)
Dolt's information_schema.processlist exposes `time` as a signed INT, not
INT UNSIGNED like older MySQL builds. sqlx is strict about the type
match — the u32 binding errored out every tick:

  error occurred while decoding column "time": mismatched types;
  Rust type `u32` (as SQL type `INT UNSIGNED`) is not compatible with
  SQL type `INT`

The query itself ran fine, but decoding aborted before we could KILL
anything — so the watchdog was a no-op on Dolt despite being scheduled
correctly. Switch the field to i64 and let the existing `time >= ?`
filter do the rest.

+++ server/crates/djinn-db/src/background/dolt_watchdog.rs
@@ -144,7 +144,12 @@ async fn run_tick(pool: &MySqlPool, kill_after_secs: u32) -> sqlx::Result<()> {
 #[derive(sqlx::FromRow)]
 struct StuckRow {
     id: u64,
-    time: u32,
+    // Dolt's `information_schema.processlist.time` is a **signed** `INT`
+    // (not `INT UNSIGNED`), unlike older MySQL builds. sqlx is strict about
+    // the type match, so an `i64` here keeps decoding working against both
+    // signed and unsigned flavors. Negative values are never expected; we
+    // clamp at the comparison site below by casting to u32.
+    time: i64,
     info: Option<String>,
 }
